Input Subsystem.
This is a first pass at implementing an input subsystem. At a high level
the interface is a method to start polling for keyboard events, and a
queue that is exposed for a consumer to read from.

Rather than actually trying to make system calls to detect keyboard
activity all input is read from STDIN. In order to do this some terminal
options are set so input isn't echo'd back to STDOUT and to allow
reading without needing to press return. A new thread is spawned to push
all input onto a queue. Some keys are represented by multiple characters
so a second thread continually reads from the queue to parse the raw
input into consumable symbols that represent the keys of a keyboard.

Because some keys are represented by multiple characters a hacky
solution of sleeping when we're in an ambigous state is used.

There are some limitations to detecting keyboard activity from STDIN.

- Only key presses can be detected. There is no way of detecting a
  key up events.

- Only a single key event can be detected at one time. For example
  if you held 'a' you would get a continous stream of :a symbols
  pushed onto the to parsed queue, if while remaining to hold 'a' you
  held 'b' the stream of :a symbols would be interupted and replaced
  with a stream of :b symbols.