Display Subsystem.
First pass at a display subsystem. The interface is a 2d co-ord system
that allows consumers to update single pixels by providing the x and y
co-ords. A render method clears the terminal screen and prints the 2d
display.

Internally the display uses a string as a buffer. The 2d co-ords are
translated into the appropriate index. The math is pretty
straightforward but a bit unusual as the string buffer by default uses a
row size of row + 1 to account for newlines at the end of each row.
Additionally a buffer maps to a given pixel implementation. A pixel is a
string in this context but can be implemented to be various lengths. The
buffer initializes the string to account for a given pixel size.

A sample pixel class is provided to demonstrate using escape sequences
to color and style characters.